+++ api/discounts-get/sample.dat
@@ -1,0 +1,3 @@
+{
+    "name": "Azure"
+}
